+++ packages/component-driver-mui-v5-test/src/examples/index.ts
@@ -1,7 +1,7 @@
-export * from './Button.examples';
+export * from './button';
 export * from './Checkbox.examples';
 export * from './Input.examples';
-export * from './rating/';
+export * from './rating';
 export * from './Select.examples';
 export * from './Slider.examples';
 export * from './Switch.examples';
